+++ src/command/crf_search.rs
@@ -145,12 +145,19 @@ async fn _run(
     input_probe: Arc<Ffprobe>,
     bar: ProgressBar,
 ) -> Result<Sample, Error> {
-    // Whether to make the 2nd iteration on the ~20%/~80% crf point instead of the min/max to
-    // improve interpolation particularly in the case the the correct crf is in the middle 60%.
-    let guess_middle_60 = max_crf.is_none() && *min_crf == DEFAULT_MIN_CRF && *min_vmaf >= 94.0;
-
-    let max_crf = max_crf.unwrap_or_else(|| args.encoder.default_max_crf());
+    let default_max_crf = args.encoder.default_max_crf();
+    let max_crf = max_crf.unwrap_or(default_max_crf);
     ensure_other!(*min_crf < max_crf, "Invalid --min-crf & --max-crf");
+
+    // Whether to make the 2nd iteration on the ~20%/~80% crf point instead of the min/max to
+    // improve interpolation by narrowing the crf range a 20% (or 30%) subrange.
+    //
+    // 20/80% is preferred to 25/75% to account for searches in the "middle" benefitting from
+    // having both bounds computed after the 2nd iteration, whereas the two edges must compute
+    // the min/max crf on the 3rd iter.
+    //
+    // If a custom crf range is being used under half the default, this 2nd cut is not needed.
+    let cut_on_iter2 = (max_crf - *min_crf) > (default_max_crf - DEFAULT_MIN_CRF) * 0.5;
 
     let crf_increment = crf_increment
         .unwrap_or_else(|| args.encoder.default_crf_increment())
@@ -235,7 +242,7 @@ async fn _run(
                     ensure_or_no_good_crf!(sample_small_enough, sample);
                     return Ok(sample);
                 }
-                None if guess_middle_60 && run == 1 && sample.q + 1 < max_q => {
+                None if cut_on_iter2 && run == 1 && sample.q + 1 < max_q => {
                     q = (sample.q as f32 * 0.4 + max_q as f32 * 0.6).round() as _;
                 }
                 None => q = max_q,
@@ -262,7 +269,7 @@ async fn _run(
                 Some(lower) => {
                     q = vmaf_lerp_q(*min_vmaf, &sample, lower);
                 }
-                None if guess_middle_60 && run == 1 && sample.q > min_q + 1 => {
+                None if cut_on_iter2 && run == 1 && sample.q > min_q + 1 => {
                     q = (sample.q as f32 * 0.4 + min_q as f32 * 0.6).round() as _;
                 }
                 None => q = min_q,