+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
-* crf-search: Tweak 2nd iteration logic that would previously pick the ~25%/75% crf point.
-  Widen to ~20%/80% and only apply if min-crf,max-crf have not been overridden and min-vmaf >= 94.
+* crf-search: Tweak 2nd iteration logic that slices the crf range at the 25% or 75% crf point.
+  - Widen to 20%/80% to account for searches of the "middle" two subranges being more optimal.
+  - Disable when using custom min/max crf ranges under half the default.
 
 # v0.7.19
 * Fix stdin handling sometimes breaking bash shells.